Adds `trafaret` parameter to DataError
This also sets the parameter where possible (and unambiguous)

The parameter might be used for better error display by accessing
original parameters of a trafaret.

Original intention was to implement nicer display for `Or` alternatives
in `trafaret_config` package. There is an [example test][1] that shows
the difference (note `if BEAUTY_ERROR:` in two test suites), note that
we not only can't make shorter errors we also don't have a good
heuristic to guess line numbers inside any of the alternative branches

[1]: https://github.com/tailhook/trafaret_config/blob/38255eb8f345acdc5cdef14c8e432e5716251e77/tests/test_alternatives.py